requirements.txt
discord.py[voice]==1.3.4
PyNaCl==1.3.0
cffi==1.14.0
pycparser==2.20
six==1.15.0
boto==2.49.0
psycopg2==2.8.5